rename bodo_plan to plan
