+++ bodo/pandas/_plan.cpp
@@ -1,4 +1,4 @@
-#include "_bodo_plan.h"
+#include "_plan.h"
 #include <utility>
 #include "_executor.h"
 #include "duckdb.hpp"
