+++ bodo/pandas/plan_optimizer.pyx
@@ -255,7 +255,7 @@ cdef extern from "duckdb/planner/operator/logical_get.hpp" namespace "duckdb" no
         pass
 
 
-cdef extern from "_bodo_plan.h" nogil:
+cdef extern from "_plan.h" nogil:
     cdef unique_ptr[CLogicalGet] make_parquet_get_node(c_string parquet_path, object arrow_schema)
     cdef unique_ptr[CLogicalComparisonJoin] make_comparison_join(unique_ptr[CLogicalOperator] lhs, unique_ptr[CLogicalOperator] rhs, CJoinType join_type, vector[int_pair] cond_vec)
     cdef unique_ptr[CLogicalOperator] optimize_plan(unique_ptr[CLogicalOperator])
